+++ ZenUnitUtilsAndAssertionTests/Components/Iteration/MetalMock/TransformerMock.h
@@ -6,5 +6,5 @@ class TransformerMock : public Metal::Mock<Transformer<T, TransformedT>>
 public:
    using TransformerFunctionType = TransformedT(*)(const T&);
    METALMOCK_NONVOID2_CONST(vector<TransformedT>, Transform, const vector<T>*, TransformerFunctionType)
-   METALMOCK_NONVOID3_CONST(vector<TransformedT>, RandomTransform, vector<T>*, TransformerFunctionType, unsigned)
+   METALMOCK_NONVOID3_CONST(vector<TransformedT>, RandomTransform, vector<T>*, TransformerFunctionType, unsigned short)
 };
